requirements.txt
streamlit
Pillow
fpdf
streamlit-image-coordinates